fix: modified torchoperator to avoid import errors when torch is not installed
